+++ src/conexionAPI.java
@@ -1,8 +1,8 @@
 
 public record conexionAPI(String base_code,
                           String target_code,
-                          int conversion_rate,
-                          int conversion_result) {
+                          double conversion_rate,
+                          double conversion_result) {
